# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CORE OEM API",
    "description": "\nThe CORE API surface — the stable, versioned interface that external\nintegrations and downstream distributions consume to embed CORE's\nconstitutional governance layer.\n\n**Stability policy:** see [ADR-087](https://github.com/DariuszNewecki/CORE/blob/main/.specs/decisions/ADR-087-oem-api-versioning-and-stability-policy.md).\n**Route classification:** see [CORE-OEM-API paper](https://github.com/DariuszNewecki/CORE/blob/main/.specs/papers/CORE-OEM-API.md).\n\nRoutes marked `internal` (or absent from this spec) are CORE-operator\nconcerns — daemon lifecycle, autonomy-loop triggers, CI gate dispatch.\nThey are subject to change without notice and not part of this contract.\n",
    "version": "2.11.0",
    "x-stability-policy": "https://github.com/DariuszNewecki/CORE/blob/main/.specs/decisions/ADR-087-oem-api-versioning-and-stability-policy.md"
  },
  "paths": {
...
    "/v1/fix/runs/{run_id}": {
      "get": {
        "tags": [
          "Fix"
        ],
        "summary": "Fetch a persisted fix run",
        "description": "Read back a fix run's persisted record by `run_id`. The response shape matches ADR-055 D2: run_id, status, fix_id, kind, write, result, error. `result` is the JSONB ActionResult payload written by the background task; null while the run is still pending or executing. Returns 404 if no run exists with that id.",
        "operationId": "get_fix_run_v1_fix_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/quality/tests": {
      "post": {
        "tags": [
          "quality"
        ],
        "summary": "Quality Tests",
        "description": "Dispatch pytest, optionally scoped to a path. Async.",
        "operationId": "quality_tests_v1_quality_tests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityTestsRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncDispatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/integrity/baseline": {
      "post": {
        "tags": [
          "Integrity"
        ],
        "summary": "Integrity Baseline",
        "description": "Create a SHA256 baseline of src/ and return path + file count.",
        "operationId": "integrity_baseline_v1_integrity_baseline_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrityBaselineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/integrity/verify": {
      "post": {
        "tags": [
          "Integrity"
        ],
        "summary": "Integrity Verify",
        "description": "Verify src/ against the named baseline and return ok / errors.",
        "operationId": "integrity_verify_v1_integrity_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrityVerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AsyncDispatchResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "href": {
            "type": "string",
            "title": "Href"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "href"
        ],
        "title": "AsyncDispatchResponse",
        "description": "Standard 202 response for all async-dispatch POST endpoints.\n\nCallers poll `href` to read the persisted run record once the\nbackground task completes."
      },
...
      "FixRunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "fix_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fix Id"
          },
          "target_files": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Files"
          },
          "write": {
            "type": "boolean",
            "title": "Write"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "requested_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By"
          },
          "requested_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "kind",
          "fix_id",
          "target_files",
          "write",
          "status",
          "requested_by",
          "requested_at",
          "started_at",
          "finished_at",
          "result",
          "error"
        ],
        "title": "FixRunResponse",
        "description": "GET /v1/fix/runs/{run_id} response shape (ADR-055 D2)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IntegrityBaselineResponse": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "files_hashed": {
            "type": "integer",
            "title": "Files Hashed"
          }
        },
        "type": "object",
        "required": [
          "label",
          "path",
          "files_hashed"
        ],
        "title": "IntegrityBaselineResponse",
        "description": "POST /v1/integrity/baseline response. `path` is repository-relative."
      },
      "IntegrityRequest": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label",
            "default": "default"
          }
        },
        "type": "object",
        "title": "IntegrityRequest",
        "description": "Body for POST /integrity/baseline and POST /integrity/verify.\n\nEmpty body is valid and equivalent to label='default'."
      },
      "IntegrityVerifyResponse": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "checked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked At"
          }
        },
        "type": "object",
        "required": [
          "label",
          "ok",
          "errors"
        ],
        "title": "IntegrityVerifyResponse",
        "description": "POST /v1/integrity/verify response."
      },
...
      "QualityTestsRequest": {
        "properties": {
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "requested_by": {
            "type": "string",
            "title": "Requested By",
            "default": "api"
          }
        },
        "type": "object",
        "title": "QualityTestsRequest",
        "description": "Body for POST /quality/tests."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}